feat: collectionView 최소 height 지정 및 리뷰 반영
