--- BBus/BBus/Station/Model/StationModel.swift
@@ -1,8 +1,0 @@
-//
-//  StationModel.swift
-//  BBus
-//
-//  Created by 김태훈 on 2021/11/01.
-//
-
-import Foundation
